[PROBABLY UNSTABLE] move configurator into main app
