+++ apps/farmbot/config/dev.exs
@@ -1,4 +1,4 @@
 use Mix.Config
-config :farmbot_configurator,
-  port: 4000,
+config :farmbot,
+  configurator_port: 4000,
   streamer_port: 4040
