+++ apps/farmbot/config/prod.exs
@@ -1,7 +1,8 @@
 use Mix.Config
-config :farmbot_configurator,
-  port: 80,
+config :farmbot,
+  configurator_port: 80,
   streamer_port: 4040
+
 config :tzdata, :data_dir, "/tmp"
 config :tzdata, :autoupdate, :disabled
 config :nerves_interim_wifi, regulatory_domain: "US" #FIXME
