+++ apps/farmbot/priv/static/.gitignore
@@ -1,0 +1,5 @@
+*
+!index.html
+!.gitignore
+!.easter_eggs.json
+!.farmbot_logo.png
