+++ apps/farmbot/priv/static/index.html
