+++ apps/farmbot/tsconfig.json
