+++ apps/farmbot/web/css/_colors.scss
