+++ apps/farmbot/web/css/_widgets.scss
