+++ apps/farmbot/web/ts/advanced_settings.tsx
