+++ apps/farmbot/web/ts/interfaces.ts
