--- apps/farmbot_configurator/priv/static/.gitignore
@@ -1,3 +1,0 @@
-*
-!index.html
-!.gitignore
