--- apps/farmbot_configurator/test/farmbot_configurator_test.exs
@@ -1,8 +1,0 @@
-defmodule Farmbot.ConfiguratorTest do
-  use ExUnit.Case
-  doctest Farmbot.Configurator
-
-  test "the truth" do
-    assert 1 + 1 == 2
-  end
-end
